+++ CHANGELOG.md
@@ -45,6 +45,7 @@ UIIN-3437.
 * Added a permission to delete records from SRS to a "Inventory: All permissions" and "Inventory: View, create, edit, delete holdings". Fixes UIIN-3434.
 * Change the order of items in the item list when dragging. Refs UIIN-3455.
 * Add missing permissions for instance view. Refs UIIN-3472.
+* Fix quick instance export. Refs UIIN-3504.
 
 ## [13.0.10](https://github.com/folio-org/ui-inventory/tree/v13.0.10) (2025-09-01)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v13.0.9...v13.0.10)